Job items: ordered strikethrough, decimal quantities, categorized picker
- To-order rows strike through and dim once marked ordered (OrderedAt now
  flows through the job detail response), and the quantity stepper locks.
- Quantities accept fractions (e.g. 1.56): DB columns widen to
  numeric(10,2), API models/DTOs and client move to decimal, displays
  trim trailing zeros. Typed minimum drops to 0.01; the - button still
  clamps at 1.
- The job-page parts picker groups items by category with collapsible
  headers (flat rows, not IsGrouped - WinUI crash), search also matches
  category, and checked selections survive searching.
- Detail-page Share button wired on the job page.

+++ WorkHub.Api/DTOs/Responses/JobResponses.cs
@@ -58,7 +58,7 @@ public class JobItemResponse
     public string? Description { get; set; }
     public string? PartNumber { get; set; }
     public string? Sku { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string ListType { get; set; } = string.Empty;
     public string Source { get; set; } = string.Empty; // "library" or "adhoc"
     public Guid? InventoryItemId { get; set; }
@@ -77,7 +77,7 @@ public class OrderLineResponse
     public string? Description { get; set; }
     public string? PartNumber { get; set; }
     public string? Sku { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public DateTime? OrderedAt { get; set; }
     public Guid JobId { get; set; }
     public string JobTitle { get; set; } = string.Empty;
