+++ WorkHub.Api/Models/JobInventory.cs
@@ -7,7 +7,7 @@ public class JobInventory
     public Job Job { get; set; } = null!;
     public Guid InventoryItemId { get; set; }
     public InventoryItem InventoryItem { get; set; } = null!;
-    public int Quantity { get; set; } = 1;
+    public decimal Quantity { get; set; } = 1; // fractional units allowed (e.g. 1.5 ft)
     public string ListType { get; set; } = string.Empty; // "used" or "to_order"
     public DateTime? OrderedAt { get; set; } // set when a "to_order" item is marked ordered
     public DateTime CreatedAt { get; set; }
